pyproject.toml
[project]
name = "fedcloud_catchall"
version = "0.1.0"
description = "Tools to deal with the configuration of the catch-all components"
readme = "README.md"
authors = [
    { name = "Enol Fernández", email = "enol.fernandez@egi.eu" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
requires-python = ">=3.14"
dependencies = [
    "atrope",
    "caso",
    "cloud-info-provider==1.0.2",
    "fixtures>=4.2.8",
    "httpx>=0.28.1",
    "keystoneauth1>=5.12.0",
    "oslo-config>=9.8.0",
    "pyjwt>=2.10.1",
    "python-dateutil>=2.9.0.post0",
    "python-glanceclient>=4.10.0",
    "python-keystoneclient>=5.7.0",
    "python-novaclient>=18.11.0",
    "python-openstackclient>=8.3.0",
    "pyyaml>=6.0.3",
    "respx>=0.22.0",
    "testtools>=2.8.2",
]

[project.scripts]
token-generator = "fedcloud_catchall.token_generator:main"
image-sync = "fedcloud_catchall.image_sync:main"
accounting = "fedcloud_catchall.accounting:main"

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"

[tool.uv.sources]
atrope = { git = "https://github.com/EGI-Federation/atrope", rev = "2.0.0" }
caso = { git = "https://github.com/IFCA-Advanced-Computing/caso", rev = "1584f0b3823cc1ee31d1338944e44a36b114a5ee" }

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]
validate = [
    "pyyaml>=6.0.3",
    "yq>=3.4.3",
]
